+++ custom/intrahealth/reports/fiche_consommation_report.py
@@ -5,7 +5,7 @@ from corehq.apps.reports.filters.dates import DatespanFilter
 from corehq.apps.reports.generic import GenericTabularReport
 from corehq.apps.reports.standard import CustomProjectReport, DatespanMixin
 from custom.intrahealth.filters import FicheLocationFilter
-from custom.intrahealth.reports import IntraHealtMixin
+from custom.intrahealth.reports.utils import IntraHealtMixin
 from custom.intrahealth.sqldata import FicheData
