+++ custom/intrahealth/reports/tableu_de_board_report.py
@@ -5,7 +5,7 @@ from corehq.apps.reports.graph_models import MultiBarChart, Axis
 from corehq.apps.reports.standard import CustomProjectReport, ProjectReportParametersMixin, DatespanMixin
 from corehq.apps.hqwebapp.decorators import use_nvd3
 from custom.intrahealth.filters import LocationFilter
-from custom.intrahealth.reports import IntraHealtMixin
+from custom.intrahealth.reports.utils import IntraHealtMixin
 from custom.intrahealth.sqldata import *
